delete unnecessary factorybean #740
