--- terasoluna-gfw-functionaltest-selenium/src/test/java/org/terasoluna/gfw/functionaltest/app/webdrivers/FirefoxDriverFactoryBean.java
@@ -1,43 +1,0 @@
-/*
- * Copyright (C) 2013-2018 NTT DATA Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-package org.terasoluna.gfw.functionaltest.app.webdrivers;
-
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.firefox.FirefoxProfile;
-import org.springframework.beans.factory.FactoryBean;
-
-public class FirefoxDriverFactoryBean implements FactoryBean<FirefoxDriver> {
-
-    @Override
-    public FirefoxDriver getObject() {
-        FirefoxProfile profile = new FirefoxProfile();
-        profile.setPreference("browser.startup.homepage_override.mstone",
-                "ignore");
-        profile.setPreference("network.proxy.type", 0);
-        return new FirefoxDriver(profile);
-    }
-
-    @Override
-    public Class<?> getObjectType() {
-        return FirefoxDriver.class;
-    }
-
-    @Override
-    public boolean isSingleton() {
-        return false;
-    }
-
-}
