+++ terasoluna-gfw-functionaltest-selenium/src/test/resources/META-INF/spring/seleniumContext.xml
@@ -53,18 +53,13 @@
   </bean>
 
   <beans profile="firefox default">
-    <bean id="webDriver"
-      class="org.terasoluna.gfw.functionaltest.app.webdrivers.FirefoxDriverFactoryBean"
+    <bean id="webDriver" class="org.terasoluna.gfw.functionaltest.app.FirefoxDriverFactoryBean"
       scope="prototype" />
   </beans>
   <beans profile="chrome">
-    <bean id="webDriver"
-      class="org.terasoluna.gfw.functionaltest.app.webdrivers.ChromeDriverFactoryBean"
-      scope="prototype" />
+    <bean id="webDriver" class="org.openqa.selenium.chrome.ChromeDriver" scope="prototype" />
   </beans>
   <beans profile="ie">
-    <bean id="webDriver"
-      class="org.terasoluna.gfw.functionaltest.app.webdrivers.FirefoxDriverFactoryBean"
-      scope="prototype" />
+    <bean id="webDriver" class="org.openqa.selenium.ie.InternetExplorerDriver" scope="prototype" />
   </beans>
 </beans>
